Update patch set 1
Patch Set 1: Code-Review-1

I have nothing against this change in particular but I think we should standardize the naming we use for OTG storage so that app developers can handle it in a standard way.

On the m7 we call it usbdisk. Not sure what the other devices call it but we should standardize it within CM.

Patch-set: 1