Update patch set 1
Patch Set 1:

I should also point out that it's probably a good idea to switch, even if we find a bug in the cpufreq and fix it.

The qcom hal + mpdecision is based on running ondemand and we've switched to interactive so it may be tuned in a way that is not optimal for us.

Patch-set: 1