Update patch set 1
Patch Set 1: Code-Review-1

I wonder if this commit was always causing crashes, and I just didn't notice.

I finally got a chance to run a build this afternoon and it seems like does not crash if the phone is plugged in but crashes fairly often when sitting (unplugged) with the screen off.

I suspect its' triggering some kind of problem with the suspend/resume logic.

Patch-set: 1