Update patch set 1
Patch Set 1: -Code-Review

Steve,

I think that the crashes being seen when adding this commit is a problem with the kernel changes that you pushed last week.

By changing the minimum frequency, from 1.2 GHz down to 300 MHz, this commit opens up the possibility of triggering many more timing issues in the kernel code.

I have been running most of the day (about 8 hours) with a kernel based off of commit:

8151af0 tty: n_smux: add separate close notifications

but with the rest of the code up-to-date.  I haven't had any crashes in that time (using the kernel's HEAD, I usually crash in 15-30 minutes).

Yesterday, I tried using the kernel @ HEAD and in the device reverting the commit that moves from ondemand to interactive (and without this commit).  Removing this commit and restoring ondemand is an alternative way of fixing the fact that the first two cores gets locked at a higher minimum speed.  This also crashed suggesting that it's not related to changing the power HAL.

Patch-set: 1