Update patch set 2
Patch Set 2: Code-Review+1

I just tried again with cpu boost enabled and it is working normally this time.  I am seeing plenty of 300 MHz time.

FYI, Steve, for testing I simply booted the phone and let it sit for a few minutes (on a charger) and then used CPU Spy to look at the time per freq.

I wonder if my phone was just really busy this morning or if it was still too early and I flashed the wrong build...

Patch-set: 2