Update patch set 1
Patch Set 1:

Several XDA users are reporting crashes with today's nightly.  When there was a build out there with my power HAL change a couple of people had corruption of userdata.  I had a bunch of errors that looked like small data corruption (had to clear data on some apps).

I haven't tried running with this change so I may be wrong, but, for kicks, try leaving your phone unplugged (you need to be able to deep sleep) sitting on your desk for an hour or two without touching it and look to see if it has crashed.

If it has crashed and the big PM changes aren't getting pushed, it might be nice to revert this for people flashing nightlies.

Patch-set: 1