Update patch set 3
Patch Set 3:

Update on the BT thing (in case it's related to the new kernel) I just noticed that when a BT device disconnects, as in I shut off my BT speakers, the wifi connection drops and then re-connects 15 seconds or so later.  Not a huge issue but I figured it was worth mentioning.

Patch-set: 3