Update patch set 1
Patch Set 1: Code-Review-1

Thanks for the bump.  I've been a little swamped outside of android lately so I haven't cleaned this up.  This commit was an example of adding a 3rd level of performance modes to recovery.

Steve opted against that commit.

This commit needs to be changed so that mode=1 goes away and mode=2 becomes mode=1.  I just haven't had a chance to do that and test it before pushing it back up.

Patch-set: 1