Update patch set 1
Patch Set 1:

Even these timings alone make it a lot more enjoyable, nice.

I am not convinced that sched is responsible for the slowdowns, because a) a simple switch to ondemand and an older version of interactive did away with 'em and b) a different  routine (hacky and with sideffects) inside interactive's timer function did the same and resulted amazing numbers/performance.

Patch-set: 1