Update patch set 1
Patch Set 1:

What we'll want to do is what we do on the other devices and lower the value to a much bigger value when the screen is off. Doing this can cause audio jitter with screen off though, so I added a power hint to deal with it. I will port those changes into the PowerHAL that Hammerhead is using.

Patch-set: 1