Update patch set 2
Patch Set 2:

True.

This feature is intended to be opt-in (https://github.com/CyanogenMod/android_frameworks_base/commit/2323852b776e1ccee5dd48f88f2f7a77d96a967a) but for whatever reason is not.

As I do not believe in this feature, I cannot be bothered to look into this. If everything works as intended, we can put this back in. Until then we have to keep this out because users are annoyed and confused.

Patch-set: 2