Update patch set 1
Patch Set 1:

Ethan: In the actual device image, see http://pastebin.com/QgbQgCsY I have to look at the github comments regarding the original commit as well.

Nick: These are not even actual apks so the script should give errors on purpose in this stage.

My modem died so updates are hard to do for me for a couple of days.

Patch-set: 1