Update patch set 1
Patch Set 1:

it works (display is turned on in B&W -or shades of grey- few times) but doesn't turn on if the phone is lifted. No particular excessive battery drain resulting from this.
Though I don't see the point of it if the screen is locked (password/PIN locked, it can say much more but that goes against the point of locking IMHO) it shows the name of the app and nothing more, an information that is well transmitted with the blinking led (bijective map  app <-> color)

Patch-set: 1
Reviewer: Kenshin <1220@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0