Update patch set 1
Patch Set 1: Code-Review+1

Well Kensin, this Ambient Display can be disabled via Settings, so if users think it is unusable for them they can just disable it. I like it because I am on home or class and when I receive a notification I can see what notification is without touch the phone, and if I think I need to open it I´ll open it, and if I think is not important will let the screen turn off again, as simple as that.

That´s  why Google added a toggle/checkbox to enable or disable it. It is in Settings->Display.

Patch-set: 1
Label: Code-Review=+1