Update patch set 2
Patch Set 2:

This change still does nothing. 

That XML needs to be included in media_codecs.xml for the ffmpeg codecs to be included.

See line 114 of your media_codecs.xml and compare it to the bacon, it is including the google codecs, it also needs to include the ffmpeg ones.

Patch-set: 2