Update patch set 1
Patch Set 1:

This is actually a good idea. Here, we have already written to overlay a lot of times in the past. From the top of my head, the latest upstream merges only conflicted trivially once in device.mk.

However, it's fine now so thanks for the heads-up, David.

Patch-set: 1