Update patch set 1
Patch Set 1:

Repro: start an activity that blocks the screen (i.e. youtube), switch to recents, let the screen turn off. After a second the display will turn on again and off but with backlight on. It took me a half year to find as I use my device "correctly", there may be other isolated usecases, don't know.

It's not specific to our device, it's a framework thing.

Patch-set: 1