Update patch set 1
Patch Set 1:

I've been thinking about bringing back the compatibility with AOSP for a while now, I just have to look into this. I suspect the solution is as simple as

    if (custom_hint)
        // never send NULL
    else
        // same as AOSP

The above is so simple that I don't know why I didn't do that in the first place.

Patch-set: 1