Update patch set 1
Patch Set 1:

> The NULL check won't protect you if the behavior of PowerManager is
 > changed such that the HAL receives an direct value rather than a
 > pointer.
 
Right, but we're counting on AOSP keeping a stable API upstream ;-) It's unlikely void *data will go away; similar approaches have been used in other types of notifiers for a long time.

 > Changing the API was a huge mistake and there are still many HALs
 > out there that don't handle hints properly because of that. So IMO
 > the changes in f/b are needed and I hope others will agree with me
 > on that.

Definitely agree there.

 > In any case, I have no saying here, it's not my tree and can't
 > really do anything.

When I encountered the soft reboots, it occurred to me that both checks are a good idea. We can count on data being a pointer, but if that changes, at least we'd know at compile time. Basically, it's protection against subtle bugs, regardless of the proposed f/base change (which is a very good idea imo).

Patch-set: 1