hammerhead: lower the slow charging threshold
The default slow charge threshold in the framework is set to 5000000
microwatts. This assumes the device will charge at 1a @ 5v.

The oem lg nexus 5 charger can only output 1.1a @ 4.8v at it's max. Then
that is further reduced to around 1a @ 4.3v by the charging driver in the
kernel. This results in the lock screen incorrectly displaying "Charging
slowly" all the time.

Prior to nougat the threshold was based on microampers and was set to
1000000, which the charger/device could easily do, resulting in the lock
screen displaying "Charging".

BUGBASH-39

Change-Id: Ib8f0636864d4ca6be536ec418cee9809f197488d