hammerhead: lower the slow charging threshold
The default slow charge threshold in the framework is set to 5000000
microwatts. This assumes the wall charger can output 1 amp @ 5 volts.
The oem lg nexus 5 charger can barely reach that at it's max, resulting
in the lockscreen displaying "Charging slowly" almost all the time.

Prior to nougat the threshold was solely based on microampers and was
set to 1000000, which the charger could easily do, resulting in the
the lockscreen displaying "Charging".

BUGBASH-39

Change-Id: Ib8f0636864d4ca6be536ec418cee9809f197488d