Update patch set 5
Patch Set 5:

This type of change might be fine for a new device, but I believe it does more harm than good at this late stage. And it's very disruptive to workflow since hh is undergoing 16.0 bring-up. It makes looking at git history a pain. This is the way aosp device trees are structured, so I prefer to keep that way.

Patch-set: 5
CC: Gerrit User 7118 <7118@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>